whatsapp_agent/requirements.txt
# Quantum Alpha WhatsApp Agent — Dependencies

# Server & async
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
python-dotenv==1.0.1

# AI
google-generativeai==0.7.2

# Market data
yfinance==0.2.40
